requirements.txt
requests
pillow
numpy
matplotlib
opencv-python
